"5:20pm. Done with the review. Let me have lunch. Then I'll take off the deep layers and try running the things. I am going to focus on getting self play to work in the linear regime. Getting things to work against the uniform agent got me enthused a bit too quickly.
5:55pm. Done with lunch. I had time to think about it.

```
        $"!policy.eval()" . $"!value.train()"
        loopw.for' (from: 0i32 nearTo: 5) fun _ =>
            inl r = train.vs_self 1024 (agent.neural.policy (2 ** -2) net) game
            $"!opt.step()" . $"!opt.zero_grad(True)"
```

I can make use of this. As the agent trains, the variance of the rewards should go down. Let me make sure that this is happening.

6pm.

```
        loopw.for' (from: 0i32 nearTo: 5) fun _ =>
            inl r = train.vs_self 1024 (agent.neural.policy (2 ** -2) net) game
            $"print(numpy.sqrt(numpy.sum(numpy.square(r))))"
            $"!opt.step()" . $"!opt.zero_grad(True)"
        $"print('***')"
```

Let me print this out. Right now I am starting to feel drowsy, so I think I'll call it a day after doing this test.

```
The learning rate is 2 ** -16.024677276611328
152.65677
154.86513
155.47684
155.7653
161.31941
***
-0.26427773
283.866
277.6438
274.71036
289.37686
290.4249
***
-0.16618304
192.65804
188.71974
189.92227
187.97025
192.26683
***
0.021637857
39.577988
39.848373
40.24383
39.347237
38.699867
***
6.425101e-05
48.65632
50.25607
50.767204
50.178272
50.097923
***
0.015966129
63.069355
64.83274
64.468765
62.38342
63.79257
***
-0.007875162
39.146137
39.592045
38.780052
39.76689
39.45663
***
-0.00067958795
38.643044
37.245388
38.856052
38.148556
38.006893
***
0.12115115
310.71246
327.5424
325.32574
335.05344
322.784
***
0.44634724
41.76303
40.27828
39.78962
39.548355
39.264843
***
0.020230703
38.916195
38.990288
37.777775
38.163925
37.862217
***
0.018818414
37.189674
38.460896
37.845165
39.23501
37.558426
***
0.007239379
38.71098
37.56866
37.57553
36.96223
35.8094
***
0.02909666
38.377716
37.64002
37.1909
38.33226
37.223938
***
-0.031111248
38.17269
36.94769
37.113136
37.559834
38.084213
***
0.06327398
37.446384
37.775227
37.852222
36.38512
38.27889
***
-0.018854445
37.051952
36.53298
38.225925
36.18501
37.061863
***
-0.025277866
37.451363
37.266434
36.875675
37.671062
35.876198
***
-0.0037651882
37.7656
37.288765
38.11274
37.320755
37.209904
***
0.0542758
36.208527
36.509212
36.56876
37.285484
36.808952
***
-0.020915568
---
0.030664062
```

Yeah, it is not optimizing properly. Forget deep nets for the time being. I should make sure that I know the linear.

This should be easy enough. Let me make sure the linear net trains.

```py
def small(intro,mid,out): return torch.nn.Linear(intro,out)
```

I'll just set it to this.

```
The learning rate is 2 ** -14.000831604003906
126.2426
116.51247
120.338
122.61144
121.98851
***
The mean reward is 0.18014717
418.58896
422.07748
418.40485
445.39203
423.61884
***
The mean reward is -0.32226026
491.75833
488.3371
490.75085
491.0419
481.97183
***
The mean reward is 0.16584054
542.3879
597.8814
614.6968
680.8916
866.55145
***
The mean reward is 2.095773
113.90235
93.17774
78.99749
71.5887
65.63136
***
The mean reward is -0.09679471
59.59608
54.150574
51.57915
50.163097
46.019787
***
The mean reward is 0.093424425
45.85101
43.055397
41.9789
41.47172
41.3642
***
The mean reward is -0.039637253
39.894928
39.48248
39.33128
38.861378
39.06383
***
The mean reward is -0.022923656
38.310844
38.402485
38.113415
37.474697
37.74313
***
The mean reward is -0.07358953
37.48532
37.093834
36.937134
37.231964
37.732674
***
The mean reward is -0.015646137
37.99493
36.49745
38.143406
38.00041
38.205994
***
The mean reward is -0.010188796
37.009064
35.833603
38.478622
37.60643
37.84985
***
The mean reward is 0.013285814
37.98934
38.27884
38.030552
38.272476
37.78488
***
The mean reward is 0.012319813
37.256893
36.904
36.599487
37.726475
37.221664
***
The mean reward is -0.04341011
37.06442
37.354923
37.104057
36.83956
38.429775
***
The mean reward is -0.058931507
37.31677
37.06984
38.08183
38.835236
38.50511
***
The mean reward is -0.06690371
37.756935
37.895844
37.39155
36.867706
38.05232
***
The mean reward is 0.0023752432
37.637085
38.2177
38.722855
37.73698
35.78962
***
The mean reward is 0.00958805
36.72282
37.316704
38.378662
36.65352
38.667374
***
The mean reward is -0.008728572
35.952
37.513783
37.63442
38.40881
37.17422
***
The mean reward is -0.029325718
---
-0.049316406
```

I have no idea what the reasonable values should be, but this is not improving at all. By the looks of it, I couldn't even guess that is training the value net. Actually, I see some movement near the beginning, but it is barely in the right direction.

6:10pm. I am going to do in depth research into the linear player tomorrow.

6:15pm. It is just so difficult to get a handle on this. My apporach is so inefficient. I have good idea, but it feels like I spent all my time going after the bad ones. I need to focus on the worth of debugging and introspection.

One thing that is really harming me here is the lack of ability to do live coding.

Human level AIs of the future will have a much more direct path between idea and program. I myself have to wade through code."