"1:20pm. Let me resume. It is time to get the loop out of the way.
1:40pm. Ah, let me take a break. I haven't had enough.

2:25pm. The implications of having such a sampling policy is astonishing.

There is in fact nothing stopping the sampling policy from incorporating hindsight information. This is what would be the case if I based it off value uncertainty. I could train the actor without ever acting on the actual policy!

Now that I've come to this point, this kind of training resembles less the vanilla RL training, and looks quite a lot like actual reasoning!

It seems a good definition of reasoning would be fast RL.

RL today works so poorly that it is difficult to imagine this being the case, but why wouldn't it be? Long term planning and reasoning is the logical endpoint in the development of RL. When RL is naive, it resembles the agent following its policy and collecting reward, but when it is advanced it resembles planning, imagination and dreaming.

The thing I will do with the sampling policy will work, and yet it is more like something the agent would simulate in its head rather than act out in the real world. In the real world, you certainly would not be allow to just act upon your curiosity. It would get you killed.

Here in simulation, it is perfectly fine.

2:45pm.

```
inl vs_self batch_size p game =
    let rec loop (l : a u64 _) =
        inl rewards : ra u64 _ = am.empty
        inl actions_indices : ra u64 _ = am.empty
        inl actions : ra u64 _ = am.empty
        inl nexts : ra u64 _ = am.empty
        l |> am.iteri fun i => function
            | Action: player_state,game_state,pid,actions',next =>
                rm.add actions_indices i
                rm.add actions (player_state,game_state,pid,actions')
                rm.add nexts next
            | Terminal: x =>
                rm.add rewards (i, x)
        inl rewards_actions =
            if 0 < length actions then
                inl cs,update = p actions
                am.generic.map2 (<|) nexts cs |> loop |> update
            else am.empty
        inl rewards_all : a _ r2 = create (length l)
        am.generic.iter2 (set rewards_all) actions_indices rewards_actions
        am.iter (fun (i,_,_,r) => set rewards_all i r) rewards
        rewards_all
    loop (am.init batch_size fun _ => game pl2_init)
```

Hah, I do not feel like it at all. Looking at the old stuff just saps my motivation.

I can't do it. I need to motivate myself. Let me turn off the computer and spend some time in bed. I'll try to focus my mind on how to do the Python parts.

I have the `control` and the `optimize` functions, but though I have a high level view of what I want to do, the motivation and the planning for the details of how I will do the Python side programming is lacking. I feel drained after the rigours of the last month.

I need to focus myself and hone my mind towards the new approach. I need to overcome my mental inertia.

2:50pm. The way to do that is to just think. I won't get motivated to do work just by working. I've dreamed about the high level, I need to focus on the lows."