"6:55pm. I am not going to play games or read novels like I usually do. Instead I am going right back to bed.
I had some insight.

```
        inl q_backups s i =
            inl q = ,(index qs i)
            inl r = if i = i' then (r -! q) /! sample +! q else q
            r, s +! r *! index policy i
```

Remember when I replaced the above update with...

```
        inl q_backups s i =
            inl q = ,(index qs i)
            inl r = if i = i' then r else q
            r, s +! r *! index policy i
```

...and was surprised that it worked well? I realized what this update means. This update is when the sample probability is 1. This should be obvious, but what would the update for the rest be when that sample probability is 0?

`(r -! q) /! sample` would go towards the negative or positive infinity, but since the chance of occurence is zero, then it should be zero. So it is not that the update I thought up on the spur of the moment was different - it is the implied sampling probability that was different - and incorrect given the circumstances.

7pm. For the uncertainty policy, instead of the L1 norm, I should use the max arg. I do not need to fiddle around here, I can just got for the most uncertain action directly.

7:05pm. And that is what I wanted to say.

The reason why I did not program today is due to the enormous feeling of fatigue. It feels like my will has been stretched to its limit. Last night, despite being dead tired, I had trouble falling asleep.

In the past month I've felt fear, but going in boldly and getting hit too many times has taken its toll on me.

The method I have right now is possibly the best idea ever. It will allow me to stabilize deep RL for good. With a roboust method, I can go beyond cherry picked examples and into the real world with it.

I am sure it will work, but it is not like this is the first time I've been enthusiastic about something. Have I ever succeeded?

I know that experience and skill is accumulated through failure, but if this algorithm fails...it is not that I am worried about failure. Rather I am tired of it. I am tired of false power.

The algorithm I implemented as a sketch is not something a novice or even an intermediate practitioner can do. In some ways, when it comes to putting together Lego blocks, I am an expert now.

But I do not want to hope. And I do not want to dream. This is the source of my depression.

Just one more time. Just a single success is all it would take to revive me.

This algorithm I am currently working on, this hard fought fragment of inspiration might be more important than the past six years of work combined. It is the reason why I shed so much of my life on programming. It will become the bedrock of my success.

Just how many times have I thought - what if I had the stability of tabular RL, but the power of deep learning? I just have to go a bit further and I will have it. What I've done here is on par of inventing the duality gap method for stabilizing GAN training.

10s were a wild decade, but in the 20s, machine learning is starting to mature.

7:15pm. All this knowledge, all this skill, just what is it good for? I do not want to get a wage slave job. I want to rise above the clouds.

I do not feel like making money anymore.

All I want is a single success that I can hold close to my heart. Being able to invent this method is my reward for so doggedly pursuing tabular CFR despite being able to skip stages and go to deep learning.

With this single success, I will be able to completely dominate the lower end of gaming. All the board games and all the betting games. Maybe even daytrading as long as I carefully pick spots not dominated by the big boys close to the servers. Bots could make a lot of money in places like penny stocks where my competition would mostly be fools.

But I do not have enough money to even open a trading account. Hell, to daytrade in the US markets I need over 25k otherwise I will be restricted to 4 daytrades per week. It might as well be a million.

I won't even consider it. Poker is where I should start. There should be plenty of fools at the lower stakes there. I can't imagine anybody going through these lengths just to win a card game.

7:30pm. Do I want to bitch and whine how the real world is too tough for me, or do I just want to rest for a while and then try again? The later.

I am going to rest and then I am going to rise again. I am going to demonstrate that I have a complete mastery of tabular RL.

7:35pm. Once I have this working stable on Leduc, I will go from being a loser to being a King.

Let me close here. I know the PL thread is up, but I'll leave it for tomorrow. I do not feel like fiddling with the review right now. Let me go back to bed."