"2pm. Let me resume.
I've really been into Thunderbolt Fantasy. I am actually 2 eps from finishing S1. Had I known it was this good, I would have picked it up back when it was airing.

Let me focus. I think I've decided on how I am going to deal with observations. I'll just group them and map them to a token observation. I'll ignore that this would put a token at the start, it does not matter. Though I can just always filter that out.

2:05pm. It is really so hard. I keep pushing against the boulder in my mind, trying to get it to move. It is not like I can't do any of the pieces, but I keep searching for the feeling of rightness. When I feel that the idea is right, I will have no problem going forward.

Let me step away from the screen. I've been doing that a lot it seems, but I feel the need to do it for a bit more.

So I have enough to start right now?

4:50pm. Let me step out for a while.

7:30pm. Done with lunch and that haircut. It seems there will be a heat wave later in the week, so that should keep me cooler. Right now the temperatures are great. There is even a cool wind outside. But in general, during summer the weather is too damn hot. This ideal balance is way too rare.

But that is not what I wanted to talk about.

I am still trying to overcome my inertia. I know why this is, and why my regrets are flaring up so much.

My heart is telling me - at this point doing the superhuman poker agent for any kind of game, the game itself and the UIs. All of that is a month work at most.

But even if I had the superhuman agent does that mean I can just start printing money? In an ideal world that would be true. But in reality, I suspect that I would get busted by the host.

Ultimately, the real challenge is not training the agent. Now that I understand backprop completely, I know how to make neural nets work.

The real challenge is passing the Turing test. The real challenge is and was always to pretend to be a normie. The challenge is to obey the unspoken rules of society.

7:35pm. I am afraid of this. This is the wound I've been carrying for six years. This is the root of my regret.

I figured out the secrets of omnipotence, but in the end that just impressed upon me how powerful the rules of society are. Big brain, and big muscles are not enough. Just the fist is not enough to rule the world of normals.

How dreadful and demanding those people are! I can't imagine beating them even if I was literally omnipotent!

7:40pm. Ok, so the rules are what they are. What is a repeated punch in the gut for me is that I keep looking back, acknowledge my mistakes, but I can also see that given my reasoning and energy budget I could not have acted any other way! For the sake of my path, the only choice I ever had was to lose and be at the bottom.

I couldn't have run away, because I did not have the programming skills back then.

So my mind is trapped in a recurrent nightmare where I keep trying to reason out how to beat the boss of the game at lvl 1. And I can't really win or take revenge, because that is not in the rules of the game. Anything I could come up with would just be fraud and sensenless violence.

It is just so dreadful. To win I'd have to be somebody else, but if I was such a person, I would never have figured out the self improvement loop.

7:45pm. And now I have this problem with poker. I'll have the superhuman agent soon enough, but it is a complete mystery whether I can win.

I can make a superhuman agent at playing poker. Or Go or Chess. Eventually that will be Dota and Starcraft.

But can I train an agent to be a normie?

7:50pm. I'll accept it as a challenge. I won't take it as an insult that I have to do this. In return I am definitely going to drain them dry. I will show them how little the rules of righteousness mean.

7:55pm. The ultimate challenge in programming for a human is to automate away the social game. If I could do this, I can consider myself a great programmer. And I will be able to put my regrets behind me, knowing that I've beaten the game.

8pm. I need to want it. I need to seek and believe in it. I need to dream about it and hope for it. I need to trust that it is possible.

8:05pm. In games far worse things happen that in real life, but you just reload and move on. I'll forever be pressured by my failures unless I figure out the solution to the game that I could not beat. If I could beat it, I could stop being obsessed by it.

Really I can understand why some people get consumed and start moralizing about social skills being a virtue and what not rather than just being another tool in the sheed. Even I myself wrote that if I do not manage anything but 2022 I'll give into normality. But I do not really mean it.

Yes, the power of escape is awesome. But what would be better is to beat the game in the way it is intended to be beated. To win what I need to do is make a proxy so good, that gives me all the benefits of playing it, but takes none of my energy.

8:10pm. This is all about me. I could make a superhuman agent, but then have to enter the cards manually. I could go a step forward and have it scan the screen and show me the results in an overlay. The step after that is to have it take control of the keyboard and the mouse.

The client will be watching what I do. That last step will be hard. Even if I spend a full year working on it, will it be enough? I am not sure.

Neurochips would help as they would allow me to reduce the energy expenditure of visual scene parsing. They might allow me to use GANs. But ultimately, whether I can beat the normie test depends on my own ingenuity.

8:15pm. The fact that I haven't been doing much programming lately, isn't it because deep down I still don't believe I can win?

Let me close here as I am tired. I will start in earnest when I gather the courage to fight.

If needed I will spend tomorrow in bed again. I will spend my day staring into the eye of terror until I find my resolve."