"9:40am. I keep going to bed too late. 2am is just not enough rest. Let me chill a bit first.
https://robotic.substack.com/p/reward-is-not-enough

Let me read this.

9:45am. There is something I need to test out on Leduc. I'll remove the value accumulation phase and update all the nets at the same time. If this works, it will speed up the training by 4x which is nothing to scoff at.

9:50am. Forget this. Let me read a chapter or two from the story from yesterday and then I will start programming.

10:25am. Let me set the thing to run and I will start work on putting the the two classes I made yesterday into Holdem.

```py
    def run(policy_lr,is_avg=False):
        nonlocal head,heada,t
        # for a in range(3):
        #     vs_self(batch_size,neural_player(True,False,False,2 ** -2))
        vs_self(batch_size,neural_player(True,True,True,2 ** -2))
```

I'll just do this. Let me try it out.

10:35am.

```
The mean is -0.0631682500243187 for the regular player.
The mean is -0.0837344080209732 for the average player.
```

The first try ended up being insanely good. This is good indication that I should in fact not have two phases. This is really good.

It means that instead of taking a month to train an agent, I can do it in a single week. That is what 4x improvement means.

10:40am. Focus me. Leave the training of this aside...

Ah yeah. I forgot about the review. Well, I'll leave it for later. Let me plug the new pieces.

10:40am. Damn it, I need to find a way to configure VS Code's Python plugin so it can find the regular imports. What I have now is ridiculous.

https://stackoverflow.com/questions/53939751/pylint-unresolved-import-error-in-visual-studio-code

Let me spend time to resolve this.

https://github.com/microsoft/python-language-server/blob/master/TROUBLESHOOTING.md#unresolved-import-warnings

> If you're getting a warning about an unresolved import, first ensure that the package is installed into your environment if it is a library (pip, pipenv, etc). If the warning is about importing your own code (and not a library), continue reading.

Seems promising.

```
{
    "python.pythonPath": "C:\\Users\\Marko\\anaconda3\\python.exe",
    "python.autoComplete.extraPaths": ["./cython_experiments/ui_holdem1"]
}
```

Let me see if it starts working now.

10:55am. I have no idea what is wrong, and annoyingly I can't restart the editor until the training finishes.

It works with a single directory, but fails when the directories are nested. Should I just open the folder directly?

```
"python.autoComplete.extraPaths": ["C:/Users/Marko/Source/Repos/The Spiral Language/Spiral Compilation Tests/cython_experiments/ui_holdem1"]
```

Let me try this.

No, it fails. Now I am stuck waiting until the training finishes before I can try restarting the server. Nevermind, let me just move forward. I need to make sure I run the training in a separate vs code instance in the future.

11:10am.

```
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.0631682500243187 for the regular player.
The mean is -0.0837344080209732 for the average player.
The mean is -0.12758806347846985 for the regular player.
The mean is -0.07584558427333832 for the average player.
The mean is -0.16126713156700134 for the regular player.
The mean is -0.07929947972297668 for the average player.
----
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.15658290684223175 for the regular player.
The mean is -0.10577544569969177 for the average player.
The mean is -0.10848427563905716 for the regular player.
The mean is -0.09473014622926712 for the average player.
The mean is -0.2177843302488327 for the regular player.
The mean is -0.10020226985216141 for the average player.
----
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.18220320343971252 for the regular player.
The mean is -0.0774112120270729 for the average player.
The mean is -0.17852811515331268 for the regular player.
The mean is -0.07282959669828415 for the average player.
The mean is -0.12132035940885544 for the regular player.
The mean is -0.073543980717659 for the average player.
----
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.10726352035999298 for the regular player.
The mean is -0.1193193569779396 for the average player.
The mean is -0.2397441267967224 for the regular player.
The mean is -0.10686831921339035 for the average player.
The mean is -0.08434759825468063 for the regular player.
The mean is -0.10087884962558746 for the average player.
----
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.15739214420318604 for the regular player.
The mean is -0.08799194544553757 for the average player.
The mean is -0.14317815005779266 for the regular player.
The mean is -0.0799565389752388 for the average player.
The mean is -0.13005416095256805 for the regular player.
The mean is -0.08690895885229111 for the average player.
```

The training is done in record time. And the performance is excellent. What this means is that I improved by a factor of over a hundred compared to the...

https://arxiv.org/abs/1810.09026
Actor-Critic Policy Optimization in Partially Observable Multiagent Environments

...above paper. Not having to take over a 100 steps to optimize the value net before every actor update will make a massive difference in compute time.

11:35am. Added a comment to the package file.

This discovery is simple, but remarkable. I can do this. Let me commit here so the comment gets glued. Then I will go back to figuring out how to make Python imports work."