"7:50pm.
```
    inl deckInit = a ;[King; Queen; Jack; King; Queen; Jack]
    $"numpy.shuffle(!deckInit)"
```

Let me do a bit extra. I've realized that the tabular player can't be related to the serializer errors since it is not using it. Instead I forgot that I changed the nodes and so the games need to be modified as well to shuffle the initial elements.

Let me try running the tabular agent again with this fix.

```
Training the tabular agent.
30
60
90
120
150
180
210
240
270
300
Averaging the tabular agent.
30
The mean reward is -0.06790248304605484 for the regular agent.
The mean reward is -0.06422210484743118 for the average agent.
60
The mean reward is -0.044129081070423126 for the regular agent.
The mean reward is -0.1021738201379776 for the average agent.
90
The mean reward is -0.0441456064581871 for the regular agent.
The mean reward is -0.07722820341587067 for the average agent.
120
The mean reward is -0.06137392297387123 for the regular agent.
The mean reward is -0.12713350355625153 for the average agent.
150
The mean reward is -0.10251571983098984 for the regular agent.
The mean reward is -0.11559867858886719 for the average agent.
----
Training the tabular agent.
30
60
90
120
150
180
210
240
270
300
Averaging the tabular agent.
30
The mean reward is 0.0029630339704453945 for the regular agent.
The mean reward is -0.11508959531784058 for the average agent.
60
The mean reward is -0.09664859622716904 for the regular agent.
The mean reward is -0.10322226583957672 for the average agent.
90
The mean reward is -0.07900533080101013 for the regular agent.
The mean reward is -0.06504924595355988 for the average agent.
120
The mean reward is -0.06545975804328918 for the regular agent.
The mean reward is -0.1292322725057602 for the average agent.
150
The mean reward is -0.08379856497049332 for the regular agent.
The mean reward is -0.12437926232814789 for the average agent.
```

Seems decent. Now I could try running the nn on this, but I'd need to remove that assert and retrain the tabular agent for that. The annoying thing is that the pickle object depends on the exact module.

Well, let me do it.

8:35pm.

```
Training the tabular agent.
30
60
90
120
150
180
210
240
270
300
Averaging the tabular agent.
30
The mean reward is -0.08168873190879822 for the regular agent.
The mean reward is -0.08410689234733582 for the average agent.
60
The mean reward is -0.11534712463617325 for the regular agent.
The mean reward is -0.07581590116024017 for the average agent.
90
The mean reward is -0.10288155823945999 for the regular agent.
The mean reward is -0.07678952813148499 for the average agent.
120
The mean reward is -0.07896001636981964 for the regular agent.
The mean reward is -0.10627631098031998 for the average agent.
150
The mean reward is -0.16607771813869476 for the regular agent.
The mean reward is -0.0899767279624939 for the average agent.
Training the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
325
350
375
400
425
450
475
500
525
550
575
600
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
Averaging the NN agent.
25
50
75
100
125
150
175
200
225
250
275
300
The mean is -0.12129360437393188 for the regular player.
The mean is -0.13796742260456085 for the average player.
The mean is -0.07606924325227737 for the regular player.
The mean is -0.09432142972946167 for the average player.
The mean is -0.1058015525341034 for the regular player.
The mean is -0.08983549475669861 for the average player.
----
```

So Leduc works much like before. I'll figure out why the serializer is acting up tomorrow.

> 'I will have a superhuman agent in 3 weeks.'
> Gives up on the quest and gets a job instead.
> I wonder what he meant by that.

8:40am. I am absolutely going to figure out why the stupid things seem to have their fold button jammed. If I absolutely can't get my current method to work on anything but a toy task, I'll give up on my current approach and study upside-down RL. It will be a big blow to my efforts, but that is life.

Maybe the method is weak. Fine. But for poker with shallow stacks, it should not be this difficult to separate out some things for the value net.

8:55pm. Tomorrow, I should try out signSGD on Leduc with small batch sizes.

I also have no choice. I am going to have to train the value net and see just what it is learning. Is it learning value weak hands negatively and strong hands positively. How is it valuing the folds.

I am going to have to look into it. I'll dump it all in a file and study it there.

I'll leave the policy net random and just train a value net that way.

9pm. To think for a moment I thought I would be moving on to transformers. Yes, this is the deep learning that I know. Success on a toy problem made me start thinking it might not be shit.

I will get through this. At least now I know how to dump data into a file. I'll print out the probabilities and values of all the actions and come to a conclussion that way.

9:05pm. If signSGD is not working I'll implement a variant of Adam that tracks the infinity norm.

Maybe the solution to my current problems will be to just pretrain the value net for a while and make sure it gives sensible results before starting on the policy net.

9:10pm. I admit, I did not think about this problem. Upside down RL would not have it, since prediction and action selection is unified under one umbrella there.

But I'd need a GAN discriminator to separate out realistic and unrealistic action rewards."