"7:25pm.
```py
class SignSGD(Optimizer):
    """
    Does a step in the direction of the sign of the gradient.
    """
    def __init__(self,params,lr=2 ** -10,momentum=0.9):
        super().__init__(params,dict(lr=lr,momentum=momentum))
```

It was stupid of me not to do it like this originally. I made sure to modify Leduc as well. Let me go back to chilling. It might be worth running Leduc once just to see how it performs with momentum.

Let me do it.

https://memory-beta.fandom.com/wiki/Ferengi_Rules_of_Acquisition

> Never have sex with the boss' sister.
> Always have sex with the boss.

This made me lol.

```
The mean is -0.1338093876838684 for the regular player.
The mean is -0.21453121304512024 for the average player.
The mean is -0.20489166676998138 for the regular player.
The mean is -0.1729835420846939 for the average player.
The mean is -0.3546217083930969 for the regular player.
The mean is -0.15186043083667755 for the average player.
```

What the hell? Was this...Ah yes. Since I am doing full back learning, keeping stale gradients around can only hurt me.

```
The mean is -0.3219892680644989 for the regular player.
The mean is -0.17883962392807007 for the average player.
The mean is -0.2640190124511719 for the regular player.
The mean is -0.14830583333969116 for the average player.
The mean is -0.1929897964000702 for the regular player.
The mean is -0.171030655503273 for the average player.
```

It was not a fluke. Let me try it once without momo.

```
The mean is -0.1361275464296341 for the regular player.
The mean is -0.11078427731990814 for the average player.
The mean is -0.111993208527565 for the regular player.
The mean is -0.0904744416475296 for the average player.
The mean is -0.11570405960083008 for the regular player.
The mean is -0.09161736071109772 for the average player.
```

This is clear enough - momentum is making things worse. Though now the question arises is whether this is because signSGD in its raw form is more suitable for full batch learning, or whether self play has something to do with it.

How complicated.

Let me get back to resting. I'll test things out thoroughly tomorrow."