"1:40pm. https://web.stanford.edu/class/aa228/reports/2018/final96.pdf > Multi-agent games like card games offer us with outlets for testing foundational reinforcement learning algorithms and techniques. Currently, there exists uncertainty regarding how Q-Learning might perform when applied to the problem of Texas Hold’em. We investigate this uncertainty first by building a multi-agent Markov Decision Process model of the game Texas Hold’em. We then evaluate the performance of Q-Learning on the model by assigning specific Q-Learning strategies to specific agents, and comparing those agents by playing them against each other and by playing them against baseline agents (agents that follow fixed policies and do not learn). We find that while Q-Learning performs better than random in some cases, it consistently performs worse than a fixed policy that elects to call on every bet, suggesting that Q-Learning alone is insufficient for the problem of Texas Hold’em. We also find that Q-Learning performs better when it extracts more salient information about the game environment.
> it consistently performs worse than a fixed policy that elects to call on every bet

I admit, I had never imagined Holdem would be this difficult.

https://www.deepstack.ai/

I guess this is the state of the art then, but Deepstack is complicated and not at all generic.

3:40pm. https://arxiv.org/abs/1912.02877
Training Agents using Upside-Down Reinforcement Learning

I am thinking how I would implement this concept and have some ideas. Let me actually read the paper.

5:45pm. I am rolling it in my head.

I've decided. I am going to implement my own brand of upside down CFR as the next step on my journey. I am not going to use the thing in the paper as it is too complicated. The main idea is that in the process of mapping the rewards to actions, the network should develop some good internal representations.

I should be able to just take that and use that for semi tabular CFR. That is what I am going to do. Unlike in the paper, what I will do is pass in the rewards only towards the end to induce the network to do something useful will the state in the lower layers. I'll pass in the state, map it through a bunch of layers, normed square, and then use that for semi tabular CFR from which I won't backprop through.

This will be extremely convenient as I will be able to unify the actor and the critic into a single network.

5:55pm. This will be a good first step into the domain of self supervised learning.

It is easy to say that I will bring in GANs and transformers, but I need to take the smallest possible step forward.

I am going to have to go from the beginning, back to Leduc, then Flop, then I will try Holdem again.

6pm. If it was that easy, somebody else would have done it. If the researchers were able to make a lot money easily by using RL agents, you bet they would not be publishing papers and doing ML work.

The allure of ML is that its appex is also the appex of programming itself. It is important to dream.

6pm. This time I am going to break through. I am reminded of my time at the end of elementary school when I won the programming national championship. I excelled, but not to the point where I could touch the real world.

The 4/5 rank in ML is waiting for me.

If I could firmly step into it, I could become a contender in this race for power.

6:15pm. It would be absolutely great if I could reach it. As filled with drudgery regular programming is, reaching its heights would be divine.

Let me do it. Tomorrow I will start with the serializer - I am going to have to allow negative values in the binary serializer, so I am going to have the empty element be something other than 0. I'll set it to the min negative value.

6:20pm. I'll get rid of the hindsight information from the value net. It is not that important, it is certanly not preventing the net from choking on harder tasks.

Since I will be using tabular CFR, I'll be able to bring back the reweighting that I know and love. I'll also make a head for policy and the average policy. That later one is a possibility now. Not having to backprop and simply learning ontop of unsupervised features is a great liberator.

6:30pm. Now that I think about it, I will be able to use the average policy to sample from once I have it. That will be cool. There was an old paper before VR MCCFR showing that it gives a benefit, but I could not make use of it in the previous method. Now it is back on the table.

Tabular RL just refuses to die, even though deep RL will superseded by self supervised methods.

6:35pm. Today was the last day that I will make use of deep RL for anything."